requirements.txt
opencv-python-headless
cvzone
numpy
tensorflow
flask
pyttsx3r